Create README.loginProject
Made a login system that let's us register and writes and reads data to and from files in C++. The project involves a user registration process by asking username and password. Upon successful registration, a user file is created with the credentials. If the user does not exist, upon login, an error will be shown.